React query code made into custom hooks
