+++ 19-The-wild-oasis/src/features/cabins/CabinRow.jsx
@@ -1,10 +1,9 @@
-import styled from "styled-components"
-import { formatCurrency } from "../../utils/helpers"
-import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { deleteCabin } from "../../services/apiCabins"
-import toast from "react-hot-toast"
 import { useState } from "react"
+import styled from "styled-components"
+
 import NewCabinForm from "./NewCabinForm"
+import { formatCurrency } from "../../utils/helpers"
+import { useDeleteCabin } from "./useDeleteCabin"
 
 const TableRow = styled.div`
   display: grid;
@@ -57,19 +56,7 @@ function CabinRow({ cabin }) {
     image,
   } = cabin
 
-  const queryClient = useQueryClient()
-
-  const { isLoading: isDeleting, mutate } = useMutation({
-    mutationFn: deleteCabin,
-    onSuccess: () => {
-      toast.success("Cabin successfully deleted")
-
-      queryClient.invalidateQueries({
-        queryKey: ["cabins"],
-      })
-    },
-    onError: (err) => toast.error(err.message),
-  })
+  const { isDeleting, deleteCabin } = useDeleteCabin()
 
   return (
     <>
@@ -78,10 +65,14 @@ function CabinRow({ cabin }) {
         <Cabin>{name}</Cabin>
         <div>Fits up to {maxCapacity} guests</div>
         <Price>{formatCurrency(regularPrice)}</Price>
-        <Discount>{formatCurrency(discount)}</Discount>
+        {discount ? (
+          <Discount>{formatCurrency(discount)}</Discount>
+        ) : (
+          <span>&mdash;</span>
+        )}
         <div>
           <button onClick={() => setShowForm((show) => !show)}>Edit</button>
-          <button onClick={() => mutate(cabinId)} disabled={isDeleting}>
+          <button onClick={() => deleteCabin(cabinId)} disabled={isDeleting}>
             Delete
           </button>
         </div>
