+++ 19-The-wild-oasis/src/features/cabins/CabinTable.jsx
@@ -1,8 +1,8 @@
 import styled from "styled-components"
-import { useQuery } from "@tanstack/react-query"
-import { getCabins } from "../../services/apiCabins"
+
 import Spinner from "../../ui/Spinner"
 import CabinRow from "./CabinRow"
+import { useCabins } from "./useCabins"
 
 const Table = styled.div`
   border: 1px solid var(--color-grey-200);
@@ -29,14 +29,7 @@ const TableHeader = styled.header`
 `
 
 function CabinTable() {
-  const {
-    isLoading,
-    data: cabins,
-    error,
-  } = useQuery({
-    queryKey: ["cabins"],
-    queryFn: getCabins,
-  })
+  const { isLoading, cabins } = useCabins()
 
   if (isLoading) return <Spinner />
