+++ 19-The-wild-oasis/src/features/cabins/NewCabinForm.jsx
@@ -1,7 +1,4 @@
-import styled from "styled-components"
-import toast from "react-hot-toast"
 import { useForm } from "react-hook-form"
-import { useMutation, useQueryClient } from "@tanstack/react-query"
 
 import Input from "../../ui/Input"
 import Form from "../../ui/Form"
@@ -10,17 +7,16 @@ import FileInput from "../../ui/FileInput"
 import Textarea from "../../ui/Textarea"
 import FormRow from "../../ui/FormRow"
 
-import { createEditCabin } from "../../services/apiCabins"
-
-const Label = styled.div`
-  font-weight: 500;
-`
+import { useCreateCabin } from "./useCreateCabin"
+import { useEditCabin } from "./useEditCabin"
 
 function NewCabinForm({ cabinToEdit = {} }) {
+  const { isCreating, createCabin } = useCreateCabin()
+  const { isEditing, editCabin } = useEditCabin()
+  const isWorking = isCreating || isEditing
+
   const { id: editId, ...editValues } = cabinToEdit
   const isEditSession = Boolean(editId)
-
-  const queryClient = useQueryClient()
 
   const {
     register,
@@ -32,38 +28,12 @@ function NewCabinForm({ cabinToEdit = {} }) {
     defaultValues: isEditSession ? editValues : {},
   })
 
-  const { mutate: createCabin, isLoading: isCreating } = useMutation({
-    mutationFn: createEditCabin,
-    onSuccess: () => {
-      toast.success("New cabin successfully created.")
-      queryClient.invalidateQueries({ queryKey: ["cabins"] })
-      reset()
-    },
-    onError: (err) => {
-      toast.error(err.message)
-    },
-  })
-
-  const { mutate: editCabin, isLoading: isEditing } = useMutation({
-    mutationFn: ({ newCabinData, id }) => createEditCabin(newCabinData, id),
-    onSuccess: () => {
-      toast.success("Cabin successfully edited.")
-      queryClient.invalidateQueries({ queryKey: ["cabins"] })
-      reset()
-    },
-    onError: (err) => {
-      toast.error(err.message)
-    },
-  })
-
-  const isWorking = isCreating || isEditing
-
   function onSubmit(data) {
     const image = typeof data.image === "string" ? data.image : data.image[0]
 
     if (isEditSession)
       editCabin({ newCabinData: { ...data, image }, id: editId })
-    else createCabin({ ...data, image })
+    else createCabin({ ...data, image }, { onSuccess: (data) => reset() })
   }
 
   function onError(errors) {
