+++ 19-The-wild-oasis/src/features/cabins/useDeleteCabin.js
@@ -1,0 +1,21 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query"
+import toast from "react-hot-toast"
+import { deleteCabin as deleteCabinApi } from "../../services/apiCabins"
+
+export function useDeleteCabin() {
+  const queryClient = useQueryClient()
+
+  const { isLoading: isDeleting, mutate: deleteCabin } = useMutation({
+    mutationFn: deleteCabinApi,
+    onSuccess: () => {
+      toast.success("Cabin successfully deleted")
+
+      queryClient.invalidateQueries({
+        queryKey: ["cabins"],
+      })
+    },
+    onError: (err) => toast.error(err.message),
+  })
+
+  return { isDeleting, deleteCabin }
+}
