+++ 19-The-wild-oasis/src/features/cabins/useEditCabin.js
@@ -1,0 +1,20 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query"
+import toast from "react-hot-toast"
+import { createEditCabin } from "../../services/apiCabins"
+
+export function useEditCabin() {
+  const queryClient = useQueryClient()
+
+  const { mutate: editCabin, isLoading: isEditing } = useMutation({
+    mutationFn: ({ newCabinData, id }) => createEditCabin(newCabinData, id),
+    onSuccess: () => {
+      toast.success("Cabin successfully edited.")
+      queryClient.invalidateQueries({ queryKey: ["cabins"] })
+    },
+    onError: (err) => {
+      toast.error(err.message)
+    },
+  })
+
+  return { isEditing, editCabin }
+}
